feat(20-06): promote static ability stubs to real handlers
- Promote 17 static ability stubs to real handlers (Indestructible,
  CantBeCountered, CantBeDestroyed, FlashBack, Shroud, Vigilance,
  Menace, Reach, Flying, Trample, Deathtouch, Lifelink, CantTap,
  CantUntap, MustBeBlocked, CantAttackAlone, CantBlockAlone)
- Wire CantBeCountered into counter effect to prevent countering
- Fix pre-existing u32/u64 type mismatch in fight.rs tests
- Add 6 new tests for promoted handlers

+++ crates/engine/src/game/effects/fight.rs
@@ -99,7 +99,7 @@ mod tests {
     ) -> ObjectId {
         let id = create_object(
             state,
-            CardId(state.next_object_id as u64),
+            CardId(state.next_object_id),
             owner,
             name.to_string(),
             Zone::Battlefield,
